Before deploy

+++ resolvers/brand.js
@@ -1,87 +1,139 @@
 const Brand = require("../models/Brand");
+const Product = require("../models/Product");
 const {
   isEmpty,
-  putError,
-  checkError,
+  updateUrl,
+  MESSAGE_RESPONSE,
+  _validate,
   imageUpload,
   imageUnlink,
-  checkToken,
-  stringTourl,
-  updateUrl,
+  putError,
+  duplicateData
 } = require("../config/helpers");
-const validate = require("../validations/brand");
-const slugify = require("slugify");
+const {
+  DELETE_FUNC,
+  GET_BY_PAGINATIONS,
+  GET_SINGLE_FUNC,
+  GET_ALL_FUNC,
+} = require("../config/api_functions");
+
+const {checkAwsFolder} = require("../config/aws");
+const fs = require("fs");
+
+var sdir = './assets/images/brand';
+// var ldir = './assets/images/brand/large';
+// var mdir = './assets/images/brand/medium';
+// var tdir = './assets/images/brand/thumbnail';
+// var odir = './assets/images/brand/original';
+
+if (!fs.existsSync(sdir)){
+  fs.mkdirSync(sdir);
+}
+// if (!fs.existsSync(ldir)){
+//   fs.mkdirSync(ldir);
+// }
+// if (!fs.existsSync(mdir)){
+//   fs.mkdirSync(mdir);
+// }
+// if (!fs.existsSync(odir)){
+//   fs.mkdirSync(odir);
+// }
+// if (!fs.existsSync(tdir)){
+//   fs.mkdirSync(tdir);
+// }
 
 module.exports = {
   Query: {
     brands: async (root, args) => {
-      try {
-        const brands = await Brand.find({});
-        return brands || [];
-      } catch (error) {
-        throw new Error("Something went wrong.");
-      }
+      return await GET_ALL_FUNC(Brand, "Brands");
+    },
+    brands_pagination: async (
+      root,
+      { limit, pageNumber, search, orderBy, order }
+    ) => {
+      let searchInFields = { name: { $regex: `${search}`, $options: "i" } };
+      return await GET_BY_PAGINATIONS(
+        limit,
+        pageNumber,
+        orderBy,
+        order,
+        searchInFields,
+        Brand,
+        "Brands"
+      );
     },
     brand: async (root, args) => {
-      try {
-        const brand = await Brand.findById(args.id);
-        if (!brand) {
-          throw putError("Brand not found");
-        }
-        return brand;
-      } catch (error) {
-        error = checkError(error);
-        throw new Error(error.custom_message);
-      }
+      return await GET_SINGLE_FUNC(args.id, Brand, "Brand");
     },
   },
   Mutation: {
     addBrand: async (root, args, { id }) => {
-      checkToken(id);
+      await checkAwsFolder('brand');
+      if (!id) {
+        return MESSAGE_RESPONSE("TOKEN_REQ", "brand", false);
+      }
       try {
-        // Check Validation
-        const errors = validate("addBrand", args);
-        if (!isEmpty(errors)) {
-          throw putError(errors);
+        if (args.brands && !args.brands.length) {
+          return {
+            message: errors,
+            success: false,
+          };
         }
-
+        // const errors = _validate(["brands"], args);
+        // if (!isEmpty(errors)) {
+        //   return {
+        //     message: errors,
+        //     success: false,
+        //   };
+        // }
         const brands = await Brand.find({});
-        let brandList = brands.map((brand) => brand.name);
-
+        let brandList = brands.map((brand) => brand.name.toLowerCase());
         let addBrands = [];
+        let duplicateBrands = "";
         for (let i in args.brands) {
           if (
-            !isEmpty(args.brands[i].name) &&
-            !~brandList.indexOf(args.brands[i].name)
+            !isEmpty(args.brands[i].name) && 
+            !brandList.includes(args.brands[i].name.toLowerCase())
           ) {
             args.brands[i].url = await updateUrl(args.brands[i].name, "Brand");
             args.brands[i].meta = { title: "", description: "", keywords: "" };
             addBrands.push(args.brands[i]);
+          } else {
+            duplicateBrands += args.brands[i].name+" "
           }
         }
-
+        if(duplicateBrands.length) return {
+          message: `${duplicateBrands} already exists`,
+          success: false,
+        };
         await Brand.insertMany(addBrands);
-        return await Brand.find({});
+        return MESSAGE_RESPONSE("AddSuccess", "Brands", true);
       } catch (error) {
-        error = checkError(error);
-        throw new Error(error.custom_message);
+        return MESSAGE_RESPONSE("CREATE_ERROR", "Brands", false);
       }
     },
     updateBrand: async (root, args, { id }) => {
-      checkToken(id);
+      await checkAwsFolder('brand');
+      if (!id) {
+        return MESSAGE_RESPONSE("TOKEN_REQ", "Brands", false);
+      }
+      if (!args.id) {
+        return MESSAGE_RESPONSE("ID_ERROR", "Brands", false);
+      }
       try {
-        // Check Validation
-        const errors = validate("updateBrand", args);
+        const errors = _validate(["name", "url"], args);
         if (!isEmpty(errors)) {
-          throw putError(errors);
+          return {
+            message: errors,
+            success: false,
+          };
         }
-
         const brand = await Brand.findById({ _id: args.id });
         if (brand) {
           if (args.updated_brand_logo) {
             let imgObject = await imageUpload(
-              args.updated_brand_logo,
-              "/assets/images/brand/"
+              args.updated_brand_logo.file,
+              "/assets/images/brand/",'Brand'
             );
             if (imgObject.success === false) {
               throw putError(imgObject.message);
@@ -91,39 +143,51 @@ module.exports = {
             }
           }
 
-          let url = await updateUrl(args.url, "Brand");
+          let url = await updateUrl(args.url, "Brand", args.id);
 
           brand.name = args.name;
           brand.url = url;
           brand.meta = args.meta;
           brand.updated = Date.now();
 
+          // console.log('BRAND',brand);
+          const duplicate = await duplicateData({name: args.name}, Brand, args.id)
+          if(duplicate) return MESSAGE_RESPONSE("DUPLICATE", "Brand Name", false)
           await brand.save();
-          return await Brand.find({});
+          return MESSAGE_RESPONSE("UpdateSuccess", "Brands", true);
         } else {
-          throw putError("Brand not exist");
+          return MESSAGE_RESPONSE("NOT_EXIST", "Brands", false);
         }
       } catch (error) {
-        error = checkError(error);
-        throw new Error(error.custom_message);
+        console.log(error)
+        return MESSAGE_RESPONSE("UPDATE_ERROR", "Brands", false);
       }
     },
     deleteBrand: async (root, args, { id }) => {
-      checkToken(id);
+      // return await DELETE_FUNC(id, args.id, Brand, "Brands");
+      if (!id) {
+        return MESSAGE_RESPONSE("TOKEN_REQ", "Brand", false);
+      }
+      if (!args.id) {
+        return MESSAGE_RESPONSE("ID_ERROR", "Brand", false);
+      }
       try {
         const brand = await Brand.findByIdAndRemove(args.id);
         if (brand) {
-          //return true;
           if (brand.brand_logo) {
             imageUnlink(brand.brand_logo);
           }
-          const brands = await Brand.find({});
-          return brands || [];
+          let _id = args.id;
+          const product = await Product.updateMany(
+            {},
+            { $unset: { brand: _id } }
+          );
+          console.log("dgdfhdgh");
+          return MESSAGE_RESPONSE("DELETE", "Brand", true);
         }
-        throw putError("Brand not exist");
+        return MESSAGE_RESPONSE("NOT_EXIST", "Brand", false);
       } catch (error) {
-        error = checkError(error);
-        throw new Error(error.custom_message);
+        return MESSAGE_RESPONSE("DELETE_ERROR", "Brand", false);
       }
     },
   },
